add new file for read and write classes
like API, AnnotationStore
