+++ package-parser/package_parser/cli/_run_api.py
@@ -1,13 +1,11 @@
-import json
 from pathlib import Path
 from typing import Optional
 
-from package_parser.cli._json_encoder import CustomEncoder
-from package_parser.cli._shared_constants import _API_KEY
 from package_parser.processing.api import get_api
-from package_parser.processing.api.model import API
 from package_parser.processing.dependencies import get_dependencies
-from package_parser.utils import ensure_file_exists
+
+from ._read_and_write_file import _write_api_dependency_file, _write_api_file
+from ._shared_constants import _API_KEY
 
 
 def _run_api_command(
@@ -24,18 +22,3 @@ def _run_api_command(
 
     if result_dict is not None:
         result_dict[_API_KEY] = api_file_path
-
-
-def _write_api_file(api: API, out_dir_path: Path) -> Path:
-    out_file_api = out_dir_path.joinpath(f"{api.package}__api.json")
-    ensure_file_exists(out_file_api)
-    with out_file_api.open("w") as f:
-        json.dump(api.to_json(), f, indent=2, cls=CustomEncoder)
-    return out_file_api
-
-
-def _write_api_dependency_file(api: API, api_dependencies, out):
-    out_file_api_dependencies = out.joinpath(f"{api.package}__api_dependencies.json")
-    ensure_file_exists(out_file_api_dependencies)
-    with out_file_api_dependencies.open("w") as f:
-        json.dump(api_dependencies.to_json(), f, indent=2, cls=CustomEncoder)
